requirements.txt
langchain==1.2.10
langchain-google-genai==4.2.0
langchain-community==0.4.1
langchain-huggingface==1.2.0
sentence-transformers==5.2.2
chromadb==1.5.0
pdfplumber==0.11.9
streamlit==1.54.0
python-dotenv==1.2.1
